+++ tests/directives/use_module_2/module.pl
@@ -9,31 +9,26 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 
-:- object(tests,
-	extends(lgtunit)).
+:- if(current_logtalk_flag(prolog_dialect, eclipse)).
 
-	:- info([
-		version is 1.0,
-		author is 'Paulo Moura',
-		date is 2014/04/28,
-		comment is 'Unit tests for the use_module/2 built-in directive.'
-	]).
+	:- module(module).
 
-	% test all possible syntaxes for the directive
-	:- use_module(module, [
-		p/1, q/1 as a1/1, r/1:a2/1
-	]).
+	:- export((p/1, q/1, r/1)).
 
-	test(use_module_2_1) :-
-		p(X),
-		X == 1.
+	p(1).
 
-	test(use_module_2_2) :-
-		a1(X),
-		X == 2.
+	q(2).
 
-	test(use_module_2_3) :-
-		a2(X),
-		X == 3.
+	r(3).
 
-:- end_object.
+:- else.
+
+	:- module(module, [p/1, q/1, r/1]).
+
+	p(1).
+
+	q(2).
+
+	r(3).
+
+:- endif.